requirements.txt
beautifulsoup4 (4.6.0)
Flask (0.12.2)
lxml (4.1.1)
pip (1.5.4)
pygraphviz (1.3.1)
PyMySQL (0.8.0)
python-telegram-bot (9.0.0)
transitions (0.6.3)
urllib3 (1.7.1)
